feat: change kLast to uint224 type and group it together with oracle index

+++ src/interfaces/IConstantProductPair.sol
@@ -22,7 +22,7 @@ interface IConstantProductPair {
     function token0() external view returns (address);
     function token1() external view returns (address);
     function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
-    function kLast() external view returns (uint);
+    function kLast() external view returns (uint224);
 
     function mint(address to) external returns (uint liquidity);
     function burn(address to) external returns (uint amount0, uint amount1);
